revert changes ft_atoi

+++ libft/src/to/ft_atoi.c
@@ -6,15 +6,15 @@
 /*   By: ale-tron <ale-tron@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/27 15:09:37 by ale-tron          #+#    #+#             */
-/*   Updated: 2024/05/04 21:14:39 by cle-tron         ###   ########.fr       */
+/*   Updated: 2024/05/04 21:29:17 by cle-tron         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
- long long	ft_atoi(const char *str)
+ int	ft_atoi(const char *str)
 {
 	int	i;
 	int	sign;
-	unsigned long long	num;
+	int	num;
 
 	i = 0;
 	sign = 1;
